third_party/WebKit/LayoutTests/fast/dom/Range/invalid-arguments.html

<!doctype html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
var range = document.createRange();
var noNode = {};
var noRange = document.body;
[
    "compareBoundaryPoints(0, noRange)",
    "comparePoint(noNode, 0)",
    "insertNode(noNode)",
    "intersectsNode(noNode)",
    "isPointInRange(noNode, 0)",
    "selectNode(noNode)",
    "selectNodeContents(noNode)",
    "setEnd(noNode, 0)",
    "setEndAfter(noNode)",
    "setEndBefore(noNode)",
    "setStart(noNode, 0)",
    "setStartAfter(noNode)",
    "setStartBefore(noNode)",
    "surroundContents(noNode)"
].forEach(function(expr)
{
    shouldThrow("range." + expr);
});
</script>
</body>
</html>
